src/Generator/requirements.txt
google-genai
python-dotenv
vertexai
duckduckgo_search